+++ templates/sections/features.twig
@@ -2,7 +2,7 @@
 {% set subtitle = entry.featuresSubtitle %}
 {% set contentList = entry.featuresContentList.all() %}
 {% set button = entry.featuresButton %}
-<section {% if sectionId is defined %} id="{{ sectionId }}" {% endif %} class="u-section u-section--features js-bg-shape-hld js-homepage-animation-hld">
+<section {% if sectionId is defined %} id="{{ sectionId }}" {% endif %} class="u-section u-section--features js-bg-shape-hld">
 	<link rel="stylesheet" type="text/css" href="{{ alias('@web/src/css/components/features.css') }}?v=21092023">
 	<div class="bg-shape bg-shape--bottom-right js-bg-shape" data-version="bottom-right">
         <img class="s1 js-s1 js-lazy" src="data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 3 2'%3E%3C/svg%3E" data-src="{{ alias('@web/src/img/bg-shape/optimized/bottom-right-1.png' )}}" alt="Background shape blurry blob">
@@ -18,8 +18,8 @@
 			{% endif %}
             <div class="features__row">
                 <div class="features__col features__col--animation">
-                    <div class="features__animation features__animation--desktop js-homepage-animation" data-rwd="desktop" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation.json?v=07112022"></div>
-                    <div class="features__animation features__animation--mobile js-homepage-animation" data-rwd="mobile" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation-mobile.json?v=07112022"></div>
+                    <div class="features__animation features__animation--desktop js-lottie-animation" data-rwd="desktop" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation.json?v=07112022"></div>
+                    <div class="features__animation features__animation--mobile js-lottie-animation" data-rwd="mobile" data-animation-path="{% if (craft.app.sites.primarySite.baseUrl == 'http://wiliot.adchitects.co/web/') %}/web{% endif %}/src/img/homepage-animation-mobile.json?v=07112022"></div>
                 </div>
                 {% for content in contentList %}
 					{% set title = content.featureContentTitle %}
@@ -52,4 +52,6 @@
 		</div>
 	</div>
 </section>
-<fscript src="{{ alias('@web/src/js/homepageLottieAnimation.min.js') }}"></fscript>
+<!-- Already declared in supply-chain-animation.twig
+	<fscript src="{{ alias('@web/src/js/lottieAnimation.min.js') }}"></fscript>
+-->
