+++ web/src/js/local_components/supplyChainLottieAnimation.js
@@ -1,54 +1,23 @@
-console.log('supplyChainLottieAnimation script lazy-loaded');
+setTimeout(() => {
+	const warehouseTargetElements = document.querySelectorAll('g[transform="matrix(1,0,0,1,380.34698486328125,251.5240020751953)"], g[transform="matrix(0.3333300054073334,0,0,0.3333300054073334,140.50054931640625,252.0009307861328)"]');
+	const transportTargetElements = document.querySelectorAll('g[transform="matrix(1,0,0,1,686.802001953125,251.5240020751953)"], g[transform="matrix(-0.3499999940395355,0,0,0.3499999940395355,554.9000244140625,338.20001220703125)"]');
+	const warehouseAnimation = document.querySelector('.js-supply-chain-lottie-animation[data-label="warehouse"]');
+	const transportAnimation = document.querySelector('.js-supply-chain-lottie-animation[data-label="transport"]');
+	console.log(warehouseTargetElements, transportTargetElements);
 
-import Lottie from 'lottie-web';
-import gsap from 'gsap';
-import { ScrollTrigger } from 'gsap/ScrollTrigger';
-gsap.registerPlugin(ScrollTrigger);
-
-const setupSupplyChainAnimation = (animation) => {
-	const rwd = animation.dataset.rwd;
-	const mediaQuery = rwd === 'mobile' ? '(max-width: 767px)' : rwd === 'desktop' ? '(min-width: 768px)' : '';
-	if (mediaQuery && window.matchMedia(mediaQuery).matches) {
-		const anim = Lottie.loadAnimation({
-			container: animation,
-			renderer: 'svg',
-			loop: true,
-			autoplay: false,
-			path: animation.dataset.animationPath,
-		});
-		gsap.to(animation, {
-			scrollTrigger: {
-				trigger: animation,
-				start: 'top 95%',
-				end: 'bottom 10%',
-				scrub: true,
-				onEnter: () => anim.play(),
-				onEnterBack: () => anim.play(),
-				onLeave: () => anim.pause(),
-				onLeaveBack: () => anim.pause(),
-			},
-		});
-		return anim;
-	}
-	return null;
-};
-
-const initializeSupplyChainAnimations = () => {
-	ScrollTrigger.refresh();
-	const supplyChainAnimationsAll = document.querySelectorAll('.js-lottie-animation');
-	let supplyChainAnimations = Array.from(supplyChainAnimationsAll).map(setupSupplyChainAnimation);
-
-	const handleResizeSupplyChain = () => {
-		supplyChainAnimations.forEach((anim, index) => {
-			if (anim) {
-				anim.destroy();
-				supplyChainAnimations[index] = null;
-			}
-		});
-		supplyChainAnimations = Array.from(supplyChainAnimationsAll).map(setupSupplyChainAnimation);
+	const addHoverEffects = (targetElements, animation) => {
+		if (targetElements.length > 0 && animation) {
+			targetElements.forEach(targetElement => {
+				targetElement.addEventListener('mouseover', function() {
+					animation.classList.add('is-active');
+				});
+				targetElement.addEventListener('mouseout', function() {
+					animation.classList.remove('is-active');
+				});
+			});
+		}
 	};
 
-	window.addEventListener('resize', handleResizeSupplyChain);
-};
-
-initializeSupplyChainAnimations();
+	addHoverEffects(warehouseTargetElements, warehouseAnimation);
+	addHoverEffects(transportTargetElements, transportAnimation);
+}, 2000);
